Moved ContentOwner class to Models
