+++ src/Tubecode/Collections/ContentOwners.php
@@ -2,7 +2,7 @@
 
 namespace Tubecode\Collections;
 
-use Tubecode\ContentOwner;
+use Tubecode\Models\ContentOwner;
 use Tubecode\Contracts\ContentOwnerInterface;
 
 class ContentOwners
